Change to allow zoom for oversized images.
The zoom function was not working properly for the full scene images or the POM images.  It turns out that this was due to the drawing of the bounding lines for the POM and image areas.  The code has been changed to only draw these when the zoom factor is 1.